VueComponentsMainPage: Rename "views" to "page" and move it outside of "components" folder, rename "StartView" to "StartPage", "FilmShortDescription" to "FilmPreview" with relate type, delete "assets" folder and move "icons" to "components",
