+++ src/components/films/FilmPreview.vue
@@ -18,7 +18,7 @@
 import Vue from 'vue';
 import { getAppSettings } from '@/utils/appSettings';
 
-export interface FilmShortDescriptionProps {
+export interface FilmPreviewProps {
     id: string;
     posterUrl?: string;
     posterAlt?: string;
@@ -29,7 +29,7 @@ export interface FilmShortDescriptionProps {
 }
 
 export default Vue.extend({
-    name: 'FilmShortDescription',
+    name: 'FilmPreview',
     props: {
         id: {
             type: String,
@@ -58,7 +58,7 @@ export default Vue.extend({
     },
     methods: {
         onClick(id: string): void {
-            this.$emit('on-click', id);
+            this.$emit('select', id);
         },
     },
 });
